+++ src/zutils.h
@@ -21,11 +21,11 @@ typedef enum {
 
 #define Z_SEQ_INIT(name, sock) \
     zmq_msg_t name; \
-    size_t name##_opt = 1; \
+    int64_t name##_opt = 1; \
     size_t name##_len = sizeof(name##_opt); \
     void * name##_sock = (sock); \
     zmq_msg_init(&name);
-    
+
 #define Z_RECV_START(name, break_stmt) \
     if(zmq_recv((name##_sock), (&name), ZMQ_NOBLOCK) < 0) { \
         if(errno == EINTR) goto name##_finish; \
@@ -36,7 +36,7 @@ typedef enum {
     } \
     SNIMPL(zmq_getsockopt((name##_sock), ZMQ_RCVMORE, &name##_opt, &name##_len)); \
     if(!name##_opt) goto name##_error;
-    
+
 #define Z_RECV_BLOCK(name) \
     if(zmq_recv((name##_sock), (&name), ZMQ_NOBLOCK) < 0) { \
         if(errno == EINTR || errno == EAGAIN) goto name##_finish; \
@@ -44,24 +44,24 @@ typedef enum {
     } \
     SNIMPL(zmq_getsockopt((name##_sock), ZMQ_RCVMORE, &name##_opt, &name##_len)); \
     if(!name##_opt) goto name##_error;
-    
+
 #define Z_RECV(name) \
     SNIMPL(zmq_recv((name##_sock), (&name), ZMQ_NOBLOCK)); \
     SNIMPL(zmq_getsockopt((name##_sock), ZMQ_RCVMORE, &name##_opt, &name##_len));
-    
+
 #define Z_RECV_NEXT(name) \
     Z_RECV(name); \
     if(!name##_opt) goto name##_error;
-    
+
 #define Z_RECV_LAST(name) \
     Z_RECV(name); \
     if(name##_opt) goto name##_error;
 
 #define Z_SEQ_FINISH(name) \
     SNIMPL(zmq_msg_close(&name));
-    
+
 #define Z_SEQ_ERROR(name) \
     SNIMPL(zmq_msg_close(&name)); \
     if(name##_opt) skip_message(name##_sock);
-    
+
 #endif // _H_ZUTILS
